add history, command confirmation for more commands

+++ omnipod-dash/src/main/java/info/nightscout/androidaps/plugins/pump/omnipod/dash/history/data/Record.kt
@@ -4,7 +4,7 @@ sealed class Record
 
 data class BolusRecord(val amout: Double, val bolusType: BolusType) : Record()
 
-data class TempBasalRecord(val duration: Long, val rate: Double) : Record()
+data class TempBasalRecord(val duration: Int, val rate: Double) : Record()
 
 enum class BolusType {
     DEFAULT, SMB
